+++ build_cli_annotations/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.2
+
+- Added `convert` to `CliOption`.
+
 ## 0.1.1
 
 - Added `valueHelp` to `CliOption`.